+++ R/fdadist.R
@@ -18,22 +18,39 @@
 #' idx <- c(1:5, 11:15, 21:25)
 #' D <- fdadist(simulated30_sub$x[idx, ], simulated30_sub$y[idx, , ])
 fdadist <- function(x, y = NULL,
-                    warping_class = c("affine", "dilation", "none", "shift", "srsf"),
-                    metric = c("l2", "pearson"),
+                    is_domain_interval = FALSE,
+                    transformation = c("identity", "srsf"),
+                    warping_class = c("none", "shift", "dilation", "affine", "bpd"),
+                    metric = c("l2", "normalized_l2", "pearson"),
                     cluster_on_phase = FALSE,
                     labels = NULL) {
   call <- rlang::call_match(defaults = TRUE)
+  callname <- rlang::call_name(call)
+  callargs <- rlang::call_args(call)
 
-  l <- format_inputs(x, y)
+  transformation <- rlang::arg_match(transformation)
+  callargs$transformation <- transformation
+
+  warping_class <- rlang::arg_match(warping_class)
+  callargs$warping_class <- warping_class
+
+  metric <- rlang::arg_match(metric)
+  callargs$metric <- metric
+
+  l <- format_inputs(x, y, is_domain_interval)
+  check_option_compatibility(
+    is_domain_interval = is_domain_interval,
+    transformation = transformation,
+    warping_class = warping_class,
+    metric = metric
+  )
+
   x <- l$x
   y <- l$y
   dims <- dim(y)
   N <- dims[1]
   L <- dims[2]
   M <- dims[3]
-
-  warping_class <- rlang::arg_match(warping_class)
-  metric <- rlang::arg_match(metric)
 
   if (warping_class == "none" && cluster_on_phase)
     cli::cli_abort("It makes no sense to cluster based on phase variability if no alignment is performed.")
@@ -47,7 +64,7 @@ fdadist <- function(x, y = NULL,
 
   .pairwise_distances <- function(index_table) {
     pb <- progressr::progressor(steps = nrow(index_table))
-    furrr::future_map2_dbl(index_table$i, index_table$j, \(i, j) {
+    future_map2_dbl(index_table$i, index_table$j, \(i, j) {
       pb()
 
       curve_pair[1, , ] <- y[i, , ]
@@ -60,8 +77,10 @@ fdadist <- function(x, y = NULL,
         curve_pair,
         n_clusters = 1,
         seeds = 1,
-        centroid_type = "medoid",
+        is_domain_interval = is_domain_interval,
+        transformation = transformation,
         warping_class = warping_class,
+        centroid_type = "medoid",
         metric = metric,
         cluster_on_phase = cluster_on_phase,
         use_verbose = FALSE,
@@ -75,8 +94,10 @@ fdadist <- function(x, y = NULL,
         curve_pair,
         seeds = 2,
         n_clusters = 1,
-        centroid_type = "medoid",
+        is_domain_interval = is_domain_interval,
+        transformation = transformation,
         warping_class = warping_class,
+        centroid_type = "medoid",
         metric = metric,
         cluster_on_phase = cluster_on_phase,
         use_verbose = FALSE,
@@ -87,7 +108,7 @@ fdadist <- function(x, y = NULL,
         out <- max(km$distances_to_center)
 
       out
-    }, .options = furrr::furrr_options(seed = TRUE))
+    })
   }
 
   d <- .pairwise_distances(index_table)
@@ -104,8 +125,8 @@ fdadist <- function(x, y = NULL,
 }
 
 linear_index <- function(n) {
-  res <- tidyr::expand_grid(i = 1:n, j = 1:n)
-  res <- subset(res, res$j > res$i)
-  # res$k <- as.integer(n * (res$i - 1) - res$i * (res$i - 1) / 2 + res$j - res$i)
-  res
+  df <- expand.grid(j = 1:n, i = 1:n)
+  df <- subset(df, df$j > df$i)
+  rownames(df) <- NULL
+  df[, 2:1]
 }